ui update

+++ lua/modules/editor/package.lua
@@ -130,7 +130,7 @@ package({
     'numToStr/FTerm.nvim',
     config = function()
         require('FTerm').setup({
-            border = 'double',
+            border = 'single',
             dimensions = {
                 height = 0.9,
                 width = 0.9,
